Add files via upload
1. Introduction
❓ What is Aerofi
Aerofit is a leading brand in the field of fitness equipment. Aerofit provides a product range including machines such as treadmills, 
bikes, gym equipment, and fitness accessories to cater to the needs of all categories of peopl
🎯 Objective:
The market research team at AeroFit wants to identify the characteristics of the target audience for each type of treadmill offered by the
company, to provide a better recommendation of the treadmills to the new customers. The team decides to investigate whether there are
differences across the product with respect to customer characteristics.
📚 About Data:
The company collected the data on individuals who purchased a treadmill from the AeroFit stores during the prior three months.
📃 Features of the dataset:
Product Purchased: KP281, KP481, or KP781
Age: In years
Gender: Male/Female
Education: In years
MaritalStatus: Single or partnered
Usage: The average number of times the customer plans to use the treadmill each week.
Income: Annual income (in $)
Fitness: Self-rated fitness on a 1-to-5 scale, where 1 is the poor shape and 5 is the excellent shap
Miles: The average number of miles the customer expects to walk/run each week
📃 Product Portfolio:
The KP281 is an entry-level treadmill that sells for $1,500.
The KP481 is for mid-level runners that sell for $1,750.
The KP781 treadmill is having advanced features that sell for $2,500.